+++ sample/src/main/java/io/noties/adapt/sample/screen/nestedrecycler/NestedRecyclerActivity.kt
@@ -11,11 +11,9 @@ import io.noties.adapt.DiffUtilDataSetChanged
 import io.noties.adapt.Item
 import io.noties.adapt.ItemGroup
 import io.noties.adapt.ItemLayoutWrapper
-import io.noties.adapt.ViewState
 import io.noties.adapt.sample.ItemGenerator
 import io.noties.adapt.sample.R
 import io.noties.adapt.sample.screen.BaseSampleActivity
-import io.noties.debug.Debug
 
 class NestedRecyclerActivity : BaseSampleActivity() {
 
@@ -37,9 +35,11 @@ class NestedRecyclerActivity : BaseSampleActivity() {
         adapt.setItems(generator.shuffle(adapt.currentItems))
     }
 
-    // temp
+    // temp, how your items are persisted is out-of-scope of adapt
+    // most likely you would want to save state to create items (so, given the same state
+    //  the same items are created)
     companion object {
-        var items: List<Item<*>>? = null
+        var persistedItems: List<Item<*>>? = null
     }
 
     private val adapt = Adapt.create(DiffUtilDataSetChanged.create())
@@ -55,35 +55,23 @@ class NestedRecyclerActivity : BaseSampleActivity() {
         recycler.adapter = adapt
 
         if (savedInstanceState != null) {
-            val viewState = savedInstanceState.getBundle("view-state");
-            Debug.i(", restore: %s, viewState: %s",
-                    ViewState.onRestoreInstanceState(recycler, viewState), viewState)
-            adapt.setItems(items)
+            adapt.setItems(persistedItems)
+            adapt.onRestoreInstanceState(savedInstanceState, "adapt")
         } else {
             addMoreItems()
         }
     }
 
-    override fun onSaveInstanceState(outState: Bundle?) {
+    override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
 
         // hm, until view is detached, its state won't be processed and saved...
         // so, we can obtain current views from recyclerView and process then manually, but...
         //  we must have adapt id to save them...
 
-        val recycler = findViewById<RecyclerView>(R.id.recycler_view)
+        adapt.onSaveInstanceState(outState, "adapt")
 
-        for (i in 0 until recycler.childCount) {
-            val view = recycler.getChildAt(i)
-            val holder = recycler.findContainingViewHolder(view) ?: continue
-            val id = adapt.getItemId(holder.adapterPosition)
-            ViewState.save(id, view)
-        }
-
-        val viewState = ViewState.onSaveInstanceState(recycler)
-        Debug.i("viewState: %s", viewState)
-        outState?.putBundle("view-state", viewState)
-        items = adapt.currentItems
+        persistedItems = adapt.currentItems
     }
 }